Separate older pspace objects into new 'param_spaces_old.py' submodule
